move firebase authentication functions into firebase-auth-methods.ts
